switch to grid-based horizontal scrolling
The first reason for switching to grid-based horizontal scroll was to
remove the visual impact of whitespace in the DOM between pages.

It seems the scroll snapping also works better this way.

There were also some quirks in the interaction between the <details>
tag in the open state and the horizontal scrolling... these quirks
seem to be gone in this grid-based approach.