wiki panels & items are now Observable modules & variables
Committing this milestone because getting this to work was
challenging.

Some key insights that were hard won:

First:

Inspector.into() returns a function! That function returns an
inspector. Therefore panelModule() needs to use this incantation:

    return Inspector.into('main')()

Those extra parens on the right are important!

Second:

The structure for plugins works as an adaptor to the conventions of
Observable Runtime's expectations for variables. We define a function
that returns a function. The inner function expects arguments that
match the specified dependencies. The outer function serves to inject
the wiki item into the inner function's lexical context.

Third:

This code uses a global variable (window.wiki) to provide the
plugins. I tried to put those into a customization of the standard
library, but all of the ways that might work don't seem to make much
sense over just putting them in a global variable.

Forth:

DOMPurify has a parameter to emit DOM instead of a string. Will have
to figure out how to apply that knowledge.